GUI: refactor/split ui.cpp into multiple files
