+++ libgui/Makefile.am
@@ -4,4 +4,4 @@ noinst_LIBRARIES=libgui.a
 INCLUDES=-I$(top_srcdir)
 
 libgui_a_CFLAGS=
-libgui_a_SOURCES=ui.h ui.cpp
+libgui_a_SOURCES=ui.h ui.cpp ui_main.cpp qtapplication.h
